Refactor imports in admin-order-table-client, customer-form, and confirmation page; streamline auth configuration and remove deprecated code
