+++ lib/actions/admin.ts
@@ -28,66 +28,3 @@ export async function getAllOrders() {
     },
   });
 }
-
-export async function getOrderById(id: string) {
-  await requireAdmin();
-
-  const validatedId = orderIdSchema.parse(id);
-
-  const order = await prisma.order.findUnique({
-    where: { id: validatedId },
-    include: { items: true, user: true, address: true },
-  });
-
-  return order;
-}
-
-export async function getOrderByOrderNr(orderNr: string) {
-  await requireAdmin();
-
-  const validatedOrderNr = orderNumberSchema.parse(orderNr);
-
-  try {
-    const order = await prisma.order.findFirst({
-      where: {
-        orderNr: validatedOrderNr,
-      },
-      include: {
-        items: true,
-        user: true,
-        address: true,
-      },
-    });
-
-    if (!order) {
-      throw new Error("Order not found");
-    }
-
-    return {
-      customer: order.user,
-      items: order.items,
-      address: order.address,
-    };
-  } catch (error) {
-    console.error("Error fetching order:", error);
-    throw new Error("Failed to fetch order");
-  }
-}
-
-export async function updateOrderStatus(orderId: string, status: OrderStatus) {
-  await requireAdmin();
-
-  const validatedOrderId = orderIdSchema.parse(orderId);
-
-  try {
-    await prisma.order.update({
-      where: { id: validatedOrderId },
-      data: { status },
-    });
-    revalidatePath("/admin/orders");
-    return { success: true };
-  } catch (error) {
-    console.error("Failed to update order status:", error);
-    throw new Error("Failed to update order status");
-  }
-}
