+++ lib/actions/index.ts
@@ -1,3 +1,4 @@
 export * from "./admin";
 export * from "./auth";
 export * from "./product";
+export * from "./order";
