+++ lib/config/app.ts
@@ -1,4 +1,3 @@
-// Environment-based configuration
 export const APP_CONFIG = {
   DATABASE_URL: process.env.DATABASE_URL || "",
